+++ src/OpenTelemetry.Exporter.Jaeger/ApacheThrift/TBaseClient.cs
@@ -73,7 +73,8 @@ namespace Thrift
 
         public void Dispose()
         {
-            Dispose(true);
+            this.Dispose(true);
+            GC.SuppressFinalize(this);
         }
 
         protected virtual void Dispose(bool disposing)