+++ src/OpenTelemetry.Exporter.Jaeger/Implementation/Int128.cs
@@ -13,9 +13,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 // </copyright>
+
 using System;
 using System.Diagnostics;
-using System.Linq;
 using System.Runtime.InteropServices;
 
 namespace OpenTelemetry.Exporter.Jaeger.Implementation