+++ src/OpenTelemetry.Instrumentation.Http/Implementation/HttpWebRequestActivitySource.netfx.cs
@@ -23,6 +23,7 @@ using System.Reflection;
 using System.Reflection.Emit;
 using System.Runtime.CompilerServices;
 using System.Text;
+using OpenTelemetry.Context.Propagation;
 using OpenTelemetry.Trace;
 
 namespace OpenTelemetry.Instrumentation.Http.Implementation
@@ -36,15 +37,17 @@ namespace OpenTelemetry.Instrumentation.Http.Implementation
     /// </remarks>
     internal static class HttpWebRequestActivitySource
     {
-        internal const string ActivitySourceName = "OpenTelemetry.HttpWebRequest";
-        internal const string ActivityName = ActivitySourceName + ".HttpRequestOut";
+        public const string ActivitySourceName = "OpenTelemetry.HttpWebRequest";
+        public const string ActivityName = ActivitySourceName + ".HttpRequestOut";
+
+        public const string RequestCustomPropertyName = "OTel.HttpWebRequest.Request";
+        public const string ResponseCustomPropertyName = "OTel.HttpWebRequest.Response";
+        public const string ExceptionCustomPropertyName = "OTel.HttpWebRequest.Exception";
 
         internal static readonly Func<HttpWebRequest, string, IEnumerable<string>> HttpWebRequestHeaderValuesGetter = (request, name) => request.Headers.GetValues(name);
         internal static readonly Action<HttpWebRequest, string, string> HttpWebRequestHeaderValuesSetter = (request, name, value) => request.Headers.Add(name, value);
 
         internal static HttpWebRequestInstrumentationOptions Options = new HttpWebRequestInstrumentationOptions();
-
-        private const string CorrelationContextHeaderName = "Correlation-Context";
 
         private static readonly Version Version = typeof(HttpWebRequestActivitySource).Assembly.GetName().Version;
         private static readonly ActivitySource WebRequestActivitySource = new ActivitySource(ActivitySourceName, Version.ToString());
@@ -98,7 +101,7 @@ namespace OpenTelemetry.Instrumentation.Http.Implementation
 
             InstrumentRequest(request, activity);
 
-            activity.SetCustomProperty("HttpWebRequest.Request", request);
+            activity.SetCustomProperty(RequestCustomPropertyName, request);
 
             if (activity.IsAllDataRequested)
             {
@@ -115,7 +118,7 @@ namespace OpenTelemetry.Instrumentation.Http.Implementation
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private static void AddResponseTags(HttpWebResponse response, Activity activity)
         {
-            activity.SetCustomProperty("HttpWebRequest.Response", response);
+            activity.SetCustomProperty(ResponseCustomPropertyName, response);
 
             if (activity.IsAllDataRequested)
             {
@@ -131,7 +134,7 @@ namespace OpenTelemetry.Instrumentation.Http.Implementation
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private static void AddExceptionTags(Exception exception, Activity activity)
         {
-            activity.SetCustomProperty("HttpWebRequest.Exception", exception);
+            activity.SetCustomProperty(ExceptionCustomPropertyName, exception);
 
             if (!activity.IsAllDataRequested)
             {
@@ -188,32 +191,11 @@ namespace OpenTelemetry.Instrumentation.Http.Implementation
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private static void InstrumentRequest(HttpWebRequest request, Activity activity)
-        {
-            Options.TextFormat.Inject(activity.Context, request, HttpWebRequestHeaderValuesSetter);
-
-            if (request.Headers.Get(CorrelationContextHeaderName) == null)
-            {
-                // we expect baggage to be empty or contain a few items
-                using IEnumerator<KeyValuePair<string, string>> e = activity.Baggage.GetEnumerator();
-
-                if (e.MoveNext())
-                {
-                    StringBuilder baggage = new StringBuilder();
-                    do
-                    {
-                        KeyValuePair<string, string> item = e.Current;
-                        baggage.Append(WebUtility.UrlEncode(item.Key)).Append('=').Append(WebUtility.UrlEncode(item.Value)).Append(',');
-                    }
-                    while (e.MoveNext());
-                    baggage.Remove(baggage.Length - 1, 1);
-                    request.Headers.Add(CorrelationContextHeaderName, baggage.ToString());
-                }
-            }
-        }
+            => Options.TextFormat.Inject(new PropagationContext(activity.Context, activity.Baggage), request, HttpWebRequestHeaderValuesSetter);
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private static bool IsRequestInstrumented(HttpWebRequest request)
-            => Options.TextFormat.IsInjected(request, HttpWebRequestHeaderValuesGetter);
+            => Options.TextFormat.Extract(default, request, HttpWebRequestHeaderValuesGetter) != default;
 
         private static void ProcessRequest(HttpWebRequest request)
         {