+++ src/OpenTelemetry.Instrumentation.AspNet/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+* Changed the default propagation to support W3C Baggage
+  ([#1048](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1048))
+  * The default ITextFormat is now `CompositePropagator(TraceContextFormat,
+    BaggageFormat)`. Baggage sent via the [W3C
+    Baggage](https://github.com/w3c/baggage/blob/master/baggage/HTTP_HEADER_FORMAT.md)
+    header will now be parsed and set on incoming Http spans.
+
 ## 0.4.0-beta.2
 
 Released 2020-07-24