+++ src/OpenTelemetry.Instrumentation.Grpc/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Grpc instrumentation will now add the raw Request object to the Activity it
+  creates
+  ([#1099](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1099))
+
 ## 0.4.0-beta.2
 
 Released 2020-07-24